requirements.txt
asgiref==3.4.1
Django==4.0.4
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.1.0
channels==3.0.3
channels-redis==3.2.0
python-decouple==3.3
daphne==3.0.1